[login] Further split-out into .cpp files
Add new setting to disable character creation

Fix password changing to use new offset and lengths
