+++ settings/default/login.lua
@@ -34,6 +34,9 @@ xi.settings.login =
     -- Allow character deletion through the lobby (true/false)
     CHARACTER_DELETION = true,
 
+    -- Allow character creation through the lobby (true/false)
+    CHARACTER_CREATION = true,
+    
     -- Number of simultaneous game sessions per IP (0 for no limit)
     LOGIN_LIMIT = 0,
